removed gallery check for setting listed flag
previous code was preventing the ability to create listed collections at the root level.

f_list is set if you are creating a new gal/clc at the root level and are in the "Listed on Website" section. All other cases the f_list flag does not need to be set.